Refactor tests for subclasses of pattern_to_token.PatternRecognizer
